Shuffle things around between files
This adds a run_web.sh script which checks for certificate files and
runs gunicorn with them as appropriate. This updates the
docker-compose.yml and Procfile accordingly. This way the Procfile in
the image is the same as it is in the repository.

This moves the "don't write pyc files" environment setting to the
Dockerfile.

This changes a bunch of relative paths to explicit paths which made it
easier for me to understand which pieces are where.