Undo silent ignore
bin/init.py would previously silently ignore errors. That makes it
harder to figure out when there are problems. For example, right now
there's some problem with the sqlalchemy code which because the errors
are ignored and the code is in a loop causes the web container to sit in
"init" stage forever.